simplify: inline footer link, move TAB_LINKS back into tab-nav
