--- packages/app/src/components/footer/footer-links.ts
@@ -1,8 +1,0 @@
-const FOOTER_RELIABILITY_LINK = {
-  href: '/reliability',
-  label: 'GPU Reliability',
-  testId: 'footer-link-reliability',
-  event: 'footer_reliability_clicked',
-} as const;
-
-export { FOOTER_RELIABILITY_LINK };
